requirements.txt

Flask-Bcrypt
Flask-Mail
Flask==2.3.6
Flask-SQLAlchemy==3.0.5
Flask-Limiter==2.11.0
PyMySQL==1.1.1
Werkzeug==2.3.6
python-dotenv==1.0.1

